deploy: An ansible based deployment for GCS
This commit adds an initial ansible based deployment for a GD2 + CSI
driver based GCS setup. More information is available in the
deploy/README.md file.

Also present along with the ansible playbooks, is a Vagrantfile that
starts a 3 VM cluster using vagrant-libvirt, and deploys Kubernetes and
GCS on the VMs.

Kubernetes is deployed using [kubespray][1]. Kubespray is added to this
repo as a sub-module. The GCS deployment playbooks also depend upon
kubespray.

Note: As of this commit, the playbooks have been only tested on the
Vagrant environment, and should be considered as not working elsewhere.

Signed-off-by: Kaushal M <kshlmster@gmail.com>