deploy: Deploy glusterd2-cluster as individual StatefulSets
Using single-replica StatefulSets for each glusterd2 pod instead of a
single DaemonSet, allows setting up of and use of pre-known hostnames as
the listen address for glusterd2. The StatefulSets are pinned to
individual nodes.

Also, the glusterd2 pods are now deployed with 'hostPath' volumes for
/var/lib/glusterd2 which allows persistence of peerid.

With the above 2 changes, glusterd2 pods survive pod restarts.

Fixes #10, #11

Signed-off-by: Kaushal M <kshlmster@gmail.com>