Use Kubernetes v1.11
Kubernetes v1.12 brings in a few changes with respect to CSI drivers
particularly with permissions and the method of driver registration, and
requires changes to the CSI driver deployment to get it to work.

With v1.12, the CSI provisioner works well and creates volumes for PVCs
with glusterfs-csi storage class, but the CSI nodeplugin is not able to
mount volumes as the driver registration isn't done correctly.

Signed-off-by: Kaushal M <kshlmster@gmail.com>