Extra root space for atomic hosts
The CentOS Atomic VM image used by Vagrant has a relatively small root
filesystem (~10GB) which can get filled up easily by Kubernetes services
and pods.

The GCS etcd pods use an 'emptyDir' volume, which is created on the root
filesystem of the Kubernetes host, to store the etcd data. If the root
filesystem gets full, the operation of the etcd pods would be affected
possibly causing the pods to fail.

This change adds an additional 20GB disk to the Vagrant VMs, and expands
the root filesystem with the additional space made available. This
should avoid issues caused by full root filesystems.