Add retries to deployment
Deployment tended to fail at 2 points:
- When getting the gd2 peers
- When adding devices

It appears that these two failure points can be avoided by adding
retries. This commit adds those retries. We retry getting peers by just
removing the (repeat) fetch of the peers as the first instance had
retries already. When adding devices, we simply try again if necessary
during the loop.

The fixes (particularly the second) are not ideal as I don't think we
have a full understanding of the nature of the transient failures.

Signed-off-by: John Strunk <jstrunk@redhat.com>