Install socat on hosts
The lack of `socat` was preventing Helm from running in the GCS Vagrant
environment. This patch installs socat on the atomic hosts. It also
fetches the kubectl config into the local directory once installation is
complete, making it possible to use kubectl on the local machine to
access the cluster:
`$ kubectl --kubeconfig=kubeconfig get nodes`

Signed-off-by: John Strunk <jstrunk@redhat.com>