Add anthill role to correct (gcs) namespace
The Role & RoleBinding were being created in the "default" namespace
instead of the "gcs" namespace. The result was that Anthill was unable
to query the running pods due to insufficient permissions of its service
account. The result was the Anthill pod in CrashLoop. This adds the
proper namespace tags to put the Role/RoleBinding into the gcs namespace
giving the operator's service account appropriate permissions.

Signed-off-by: John Strunk <jstrunk@redhat.com>